requirements.txt
nltk
speechrecognition
pyaudio
pillow